+++ src/next/jdbc/optional.clj
@@ -91,3 +91,41 @@
   and nil columns omitted."
   [rs opts]
   (as-unqualified-modified-maps rs (assoc opts :label-fn lower-case)))
+
+(defn as-maps-adapter
+  "Given a map builder function (e.g., `as-lower-maps`) and a column reading
+  function, return a new builder function that uses that column reading
+  function instead of `.getObject` so you can override the default behavior.
+
+  This adapter omits SQL NULL values.
+
+  The default column-reader behavior would be equivalent to:
+
+      (defn default-column-reader
+        [^ResultSet rs ^ResultSetMetaData rsmeta ^Integer i]
+        (.getObject rs i))
+
+  Your column-reader can use the result set metadata to determine whether
+  to call `.getObject` or some other method to read the column's value.
+
+  `read-column-by-index` is still called on the result of that read, if
+  it is not `nil`."
+  [builder-fn column-reader]
+  (fn [rs opts]
+    (let [mrsb (builder-fn rs opts)]
+      (reify
+        rs/RowBuilder
+        (->row [this] (rs/->row mrsb))
+        (column-count [this] (rs/column-count mrsb))
+        (with-column [this row i]
+          (let [v (column-reader rs (:rsmeta mrsb) i)]
+            (if (nil? v)
+              row
+              (assoc! row
+                      (nth (:cols mrsb) (dec i))
+                      (rs/read-column-by-index v (:rsmeta mrsb) i)))))
+        (row! [this row] (rs/row! mrsb row))
+        rs/ResultSetBuilder
+        (->rs [this] (rs/->rs mrsb))
+        (with-row [this mrs row] (rs/with-row mrsb mrs row))
+        (rs! [this mrs] (rs/rs! mrsb mrs))))))